+++ src/extensions/as_egui_vec2.rs
@@ -5,7 +5,7 @@ pub trait AsEguiVec2 {
 
 impl AsEguiVec2 for Vec2 {
     fn as_egui_vec2(&self) -> egui::Vec2 {
-        egui::Vec2::new(self.x as f32, self.y as f32)
+        egui::Vec2::new(self.x, self.y)
     }
 }
