+++ src/extensions/as_vec2.rs
@@ -5,12 +5,12 @@ pub trait AsVec2 {
 
 impl AsVec2 for egui::Pos2 {
     fn as_vec2(&self) -> Vec2 {
-        Vec2::new(self.x as f32, self.y as f32)
+        Vec2::new(self.x, self.y)
     }
 }
 
 impl AsVec2 for egui::Vec2 {
     fn as_vec2(&self) -> Vec2 {
-        Vec2::new(self.x as f32, self.y as f32)
+        Vec2::new(self.x, self.y)
     }
 }
